Create driver file to separate logic
