+++ fgraph.cpp
@@ -73,21 +73,3 @@ node * fgraph::node_create(std::string line)
 
 	return new_node;
 }
-
-int main(int argc, char **argv) 
-{
-	fgraph f;
-    std::string filename;
-    std::map <int, node*> graph;
-
-	if (argc < 2) { 
-        std::cout << "Usage: fgraph <filename>" << std::endl;
-		return 1;
-	}
-	filename = argv[1];
-
-	graph = f.graph_create( f.file_read(filename) );
-	f.graph_print(graph);
-
-	return 0;
-}
